requirements.txt
streamlit
python-pptx
pandas
requests
python-dotenv
openpyxl
PyYAML
plotly
